pyproject.toml
[project]
name = "crypto-prices"
version = "0.1.0"
description = ""
authors = [
    {name = "Nima",email = "nima@newscord.org"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "google-cloud-bigquery[pandas] (>=3.36.0,<4.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "plotly (>=6.3.0,<7.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "db-dtypes (>=1.4.3,<2.0.0)",
    "anywidget (>=0.9.18,<0.10.0)",
    "nbformat (>=5.10.4,<6.0.0)"
]

[tool.poetry]
packages = [{include = "crypto_prices", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
